requirements.txt
requests==2.28.1
web3==6.5.0
eth-account==0.8.0
python-dotenv==0.21.0
numpy==1.23.3